requirements.txt
asgiref==3.5.0
distlib==0.3.4
Django==4.0.1
django-ckeditor==6.2.0
django-js-asset==2.0.0
djangorestframework==3.13.1
filelock==3.4.2
Pillow==9.0.1
platformdirs==2.4.1
pytz==2021.3
six==1.16.0
sqlparse==0.4.2
tzdata==2021.5
virtualenv==20.13.0
